+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+  * Forked from infinite list view.
+  * Fixed item counts for negative and positive
+  * Scroll boundaries can be set manually
+
 ## 1.0.1+1
   * Added New Flutter Community Banner.